+++ type-ec/src/components/app/Footer.vue
