+++ type-ec/src/components/app/Loading.vue
