+++ type-ec/src/components/itemDetail/ItemDetailContent.vue
@@ -1,0 +1,12 @@
+<template>
+<div>
+ <h1>hello</h1>
+</div>
+</template>
+
+<script lang="ts">
+import { Component, Vue,Prop} from 'vue-property-decorator';
+
+export default class ItemDetailContent extends Vue{
+}
+</script>
